% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_board}
\alias{add_board}
\title{Add board}
\usage{
add_board(name, body = NULL, token, verbose = FALSE, ...)
}
\arguments{
\item{name}{Name of the board}

\item{body}{A named list of query paramters (will be passed as body)}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST board.
}
