% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_card}
\alias{add_card}
\title{Add card}
\usage{
add_card(list, body = list(name = "New card"), ...)
}
\arguments{
\item{list}{List id.}

\item{body}{A named list of query paramters.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add card to a list.
}
