% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_card_attachment}
\alias{add_card_attachment}
\title{Add card attachment}
\usage{
add_card_attachment(
  card,
  file = NULL,
  url = NULL,
  cover = FALSE,
  name = NULL,
  ...
)
}
\arguments{
\item{card}{Card id.}

\item{file, url}{Path to a file to be attached, or a URL.}

\item{cover}{Whether the attached file should be set as cover.}

\item{name}{Name of the attachment, shown inside the card.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add attachment to a card.
}
\seealso{
Other functions to create resources: 
\code{\link{add_board}()},
\code{\link{add_card}()},
\code{\link{add_checkitem}()},
\code{\link{add_checklist}()},
\code{\link{add_comment}()},
\code{\link{add_label}()},
\code{\link{add_list}()},
\code{\link{add_member}()},
\code{\link{create_resource}()}
}
\concept{functions to create resources}
