% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_checkitem}
\alias{add_checkitem}
\title{Add checklist item}
\usage{
add_checkitem(checklist, name, checked = FALSE, position = "bottom", token,
  verbose = FALSE, ...)
}
\arguments{
\item{checklist}{Checklist id}

\item{name}{Item name (text)}

\item{checked}{Whether item should be checked; defaults to \code{FALSE}}

\item{position}{Position in the checklist; defaults to \code{"bottom"}}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST item to a checklist.
}
