% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_checklist}
\alias{add_checklist}
\title{Add checklist}
\usage{
add_checklist(card, name, source = NULL, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{name}{Checklist name}

\item{source}{Items from this checklist id will be copied to the new one (defaults to \code{NULL})}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST checklist to a card.
}
