% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_comment}
\alias{add_comment}
\title{Add comment}
\usage{
add_comment(card, text, ...)
}
\arguments{
\item{card}{Card id}

\item{text}{Comment text}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
POST comment to a card.
}
