% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{add_field_option}
\alias{add_field_option}
\title{Add dropdown option}
\usage{
add_field_option(id, text, color = "none", position = "bottom", ...)
}
\arguments{
\item{id}{Custom field ID (see \code{\link[=get_board_fields]{get_board_fields()}})}

\item{text}{Option text}

\item{color}{Option color}

\item{position}{Option position}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}}
}
\description{
Add dropdown (custom field) option.
}
\seealso{
Other functions to handle custom fields: 
\code{\link{add_field}()},
\code{\link{delete_field_option}()},
\code{\link{delete_field}()},
\code{\link{update_card_field}()},
\code{\link{update_field}()}
}
\concept{functions to handle custom fields}
