% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_label}
\alias{add_label}
\title{Add label}
\usage{
add_label(card, color, name = NULL, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{color}{Label color}

\item{name}{Label name; choosing different non-existing name will create new label (defaults to \code{NULL})}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST label to a card.
}
