% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_list}
\alias{add_list}
\title{Add list}
\usage{
add_list(board, name, position = NULL, ...)
}
\arguments{
\item{board}{Board id.}

\item{name}{List name.}

\item{position}{List position. One of \code{"top"}, \code{"bottom"} or \code{NULL}.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add list to a board.
}
