% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_member}
\alias{add_member}
\title{Add member}
\usage{
add_member(card, member, ...)
}
\arguments{
\item{card}{Card id}

\item{member}{Member id}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
POST member to a card.
}
