% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_POSIXct_hex}
\alias{as_POSIXct_hex}
\title{Convert hex string into POSIXct}
\usage{
as_POSIXct_hex(x)
}
\arguments{
\item{x}{Vector of strings, each 8 characters long}
}
\description{
Convert hex string into POSIXct
}
