% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_trello_auth_infos}
\alias{check_trello_auth_infos}
\title{Check Trello credentials}
\usage{
check_trello_auth_infos(key = NULL, token = NULL, null_return = FALSE)
}
\value{
A list including \code{api_key} and \code{api_token}.
}
\description{
This function checks if Trello's credentials exist on that local machine.
}
\details{
If a character string is given to the arguments \code{key} and \code{token}, that character string is returned as a list.
If these arguments are \code{NULL}, check if they are registered on the local machine.
If you check and the credentials do not exist, it will return an error and stop.
At this time, if the argument \code{null_return} is \code{TRUE}, no error is returned and it returns as \code{NULL}.
}
