% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_wrappers.R
\name{delete_card}
\alias{delete_card}
\title{Delete card}
\usage{
delete_card(card, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{DELETE}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{DELETE}}}
}
\description{
DELETE card.
}
