% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_wrappers.R
\name{delete_checkitem}
\alias{delete_checkitem}
\title{Delete item}
\usage{
delete_checkitem(checklist, checkitem, ...)
}
\arguments{
\item{checklist}{Checklist id}

\item{checkitem}{Checklist item id}

\item{...}{Additional arguments passed to \code{\link[=delete_resource]{delete_resource()}}}
}
\description{
DELETE checklist item.
}
