% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{delete_field_option}
\alias{delete_field_option}
\title{Delete dropdown option}
\usage{
delete_field_option(id, option, ...)
}
\arguments{
\item{id}{Custom field ID (see \code{\link{get_board_fields}})}

\item{option}{Dropdown option ID}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
Delete dropdown (custom field) option - this will remove it from all cards
on the board.
}
\seealso{
\code{\link{post_model}}, \code{\link{get_board_fields}}
}
