% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_model.R
\name{delete_model}
\alias{delete_model}
\title{DELETE data via Trello API}
\usage{
delete_model(model, id = NULL, path = NULL, token, verbose = FALSE,
  response = "content", on.error = "warning", ...)
}
\arguments{
\item{model}{Model}

\item{id}{Model id}

\item{path}{Path}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{DELETE}}}

\item{response}{Can return \code{"content"} (default), \code{"headers"} or the complete \code{"response"}}

\item{on.error}{Issues either \code{\link[base]{warning}} (default), \code{\link[base]{message}} or error (and \code{\link[base]{stop}}s)}

\item{...}{Additional arguments passed to \code{\link[httr]{DELETE}}}
}
\description{
Issues \code{\link[httr]{DELETE}} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to delete models.
}
