% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Deprecated}
\alias{Deprecated}
\alias{trello_get}
\alias{get_model}
\alias{post_model}
\alias{put_model}
\alias{delete_model}
\alias{trello_get_token}
\alias{trello_search}
\alias{search_model}
\alias{trello_search_actions}
\alias{trello_search_cards}
\alias{trello_search_boards}
\alias{trello_search_members}
\alias{trello_search_teams}
\title{Deprecated functions}
\usage{
trello_get(...)

get_model(...)

post_model(...)

put_model(...)

delete_model(...)

trello_get_token(...)

trello_search(...)

search_model(...)

trello_search_actions(...)

trello_search_cards(...)

trello_search_boards(...)

trello_search_members(...)

trello_search_teams(...)
}
\arguments{
\item{...}{See current functions for argument names and defaults.}
}
\description{
\itemize{
\item Use \code{\link[=get_resource]{get_resource()}} instead of \code{get_model()} or \code{trello_get()}
\item Use \code{\link[=update_resource]{update_resource()}} instead of \code{put_model()}
\item Use \code{\link[=create_resource]{create_resource()}} instead of \code{post_model()}
\item Use \code{\link[=delete_resource]{delete_resource()}} instead of \code{delete_model()}
\item Use \code{\link[=get_token]{get_token()}} instead of \code{trello_get_token()}
\item Use \code{\link[=search_resource]{search_resource()}} instead of \code{trello_search()}
}
}
