% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_id}
\alias{extract_id}
\title{Exctract ID}
\usage{
extract_id(x)
}
\arguments{
\item{x}{character vector of length 1}
}
\description{
Exctract model ID from model URL. If input is not a valid URL, it is returned
as is.
}
