% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board}
\alias{get_board}
\title{Get Trello Board}
\usage{
get_board(url, token = NULL)
}
\arguments{
\item{url}{url of the Trello board, or a local file downloaded via Trello JSON export}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
Read in data either via API using oauth1 token or from a local file (which you can get by downloading data from Trello and saving as a text file)
}
\examples{
# Source credentials from non-shared location (important!)
url = "url_of_your_trello_board"
b = get_board(url, token)
}

