% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board_actions}
\alias{get_board_actions}
\title{Get Board Actions}
\usage{
get_board_actions(id, token = NULL, filter = NULL, paging = FALSE,
  fix = TRUE)
}
\arguments{
\item{id}{board id}

\item{token}{previously generated token, see \code{\link{trello_get_token}} for how to obtain it}

\item{filter}{limit to just one type of action, e.g. "addMemberToCard"}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{fix}{drop and rename some columns or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with actions-related data.
}
\seealso{
\code{\link{trello_get_token}}, \code{\link{trello_get}}
}

