% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board_requests.R
\name{get_board_cards}
\alias{get_board_cards}
\title{Get All Trello Board Cards}
\usage{
get_board_cards(boardid, token)
}
\arguments{
\item{boardid}{id of the desired board}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
Returns a flat \code{data.frame} with all cards from a given Trello board.
}
\examples{
cards = get_board_cards(url, token)
}

