% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_board_infos}
\alias{get_board_infos}
\title{GET board's infos}
\usage{
get_board_infos(
  base_url,
  board_id,
  post_path = NULL,
  query = list(),
  verbose = FALSE,
  content = TRUE,
  key = NULL,
  token = NULL
)
}
\description{
WIP.
}
