% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{get_board_lists}
\alias{get_board_lists}
\title{Get All Trello Board Lists}
\usage{
get_board_lists(id, token, filter = "all", paginate = FALSE,
  simplify = TRUE)
}
\arguments{
\item{id}{id of the desired board}

\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}

\item{filter}{character whether to return only archived ("closed"), "open" or "all" (which is the default)}

\item{paginate}{logical whether to use pagination (for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{logical drop and rename some columns to make the result simpler or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Returns a flat \code{data.frame} with all lists from a given Trello board.
}
\examples{
cards = get_board_lists(url, token)
}

