% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_card.R
\name{get_card_actions}
\alias{get_card_actions}
\title{Get Card Actions}
\usage{
get_card_actions(id, ...)
}
\arguments{
\item{id}{card id}

\item{...}{additional arguments passed to \code{\link{trello_get}}}
}
\description{
Given a card ID, returns a flat \code{data.frame} with actions-related data.
}
\seealso{
\code{\link{trello_get}}
}

