% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Get All Comments From Trello Board}
\usage{
get_comments(id, token)
}
\arguments{
\item{id}{id of...}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
Complete record can only be obtained by API (due to the limit of 1000 actions), so you will need to use get_token to get it.
}
\examples{
cardid = "..."
get_comments(cardid, token)
}

