% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_flat}
\alias{get_flat}
\title{Get Data As A Flattened Data.frame}
\usage{
get_flat(url, token, query = NULL)
}
\arguments{
\item{url}{character query url}

\item{token}{previously generated token (see ?get_token for help)}

\item{query}{a list of additional url parameters such as filter = "open"}
}
\description{
This function uses http GET to download JSON via API, and returns a flat data.frame
}
\examples{
url = "https://api.trello.com/1/board/56d73a62e690ccd8d46fe5c6/cards"
all_cards = get_flat(url, token)
}

