% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\alias{get_id_board}
\alias{get_id_card}
\title{Get ID}
\usage{
get_id_board(url)

get_id_card(url)
}
\arguments{
\item{url}{Complete url, short url or just the url ID part of a Trello board}
}
\description{
Get ID of a resource.
}
\examples{
# Get Trello Development Roadmap board ID
url    = "https://trello.com/b/nC8QJJoZ/trello-development-roadmap"
tdr_id = get_board_id(url)

# Also works:
url    = "nC8QJJoZ"
tdr_id = get_board_id(url)
}

