% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_member_boards}
\alias{get_member_boards}
\title{Get Member's Boards}
\usage{
get_member_boards(id, token, paging = FALSE, fix = TRUE)
}
\arguments{
\item{id}{member ID}

\item{token}{previously generated token, see \code{\link{trello_get_token}} for info on how to obtain it}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{fix}{drop and rename some columns or return it as is. Defaults to \code{TRUE}}
}
\description{
Returns a flat \code{data.frame} with all your boards.
}

