% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_member_fields}
\alias{get_member_fields}
\title{Get Member Info}
\usage{
get_member_fields(id, fields = c("fullName", "username", "memberType", "bio"),
  ...)
}
\arguments{
\item{id}{member ID or username}

\item{fields}{by default fetches fullName, username, memberType, bio}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a flat \code{data.frame} with member-related data.
}
\seealso{
\code{\link{trello_get}}
}

