% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_member_infos}
\alias{get_member_infos}
\title{GET member's infos}
\usage{
get_member_infos(
  base_url,
  member = "me",
  post_path = NULL,
  query = list(),
  verbose = FALSE,
  content = TRUE,
  key = NULL,
  token = NULL
)
}
\description{
WIP.
}
