% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_members.R
\name{get_members}
\alias{get_members}
\title{Overview of member data}
\usage{
get_members(id, what = "boards", token)
}
\arguments{
\item{id}{member's id}

\item{what}{accepted values are 'boards' (defualt), 'organizations', 'cards'}

\item{token}{previously generated token (see ?get_token for how to get one)}
}
\description{
Get member's organizations, boards, cards via API using an oauth1 token (see ?get_token for how to get one)
}
\examples{
# Source credentials from non-shared location (important!)
source("mykeys.R")
token = get_token(key, secret)

# You will be prompted to confirm the authorization in a browser. You will also be offered an option to store the authentication in your working directory, in a hidden '.httr-oauth' file (do NOT share it with anyone!).
}

