% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pages.R
\name{get_pages}
\alias{get_pages}
\title{Get Url With Paging}
\usage{
get_pages(url, token, query = NULL)
}
\arguments{
\item{url}{Url to get}

\item{token}{Peviously generated secure token, see \code{\link{get_token}} for how to obtain it}

\item{query}{Key-value pairs which form the query, see \code{\link[httr]{GET}} for details}
}
\description{
Deprecated. Use \code{\link{paginate}}
}
\seealso{
\code{\link{get_model}}
}
