% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Request Data From Trello}
\usage{
get_request(url, token, query = list(limit = "1000"), paginate = FALSE)
}
\arguments{
\item{url}{url for the GET request, see \code{\link[httr]{GET}}}

\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}

\item{query}{url query see \code{\link[httr]{GET}}}

\item{paginate}{logical whether pagination should be used (if not, results will be limited to 50 rows)}
}
\description{
This is a generic function to perform query via trello API. Other functions such as 'get_board_cards' or 'get_board_lists' are convenience wrappers for this function.
}
\examples{
# First, build an URL including query parameters (in this case, filter = open)
id  = "id_of_a_given_board"
url = paste0("https://api.trello.com/1/boards/", id, "/cards?filter=open")

# Obtain a secure token to communicate with Trello API
token = get_token("your_key", "your_secret")

# Get all cards that are not archived
all_open = get_request(url, token)
}

