% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_org.R
\name{get_team}
\alias{get_team}
\alias{get_team_boards}
\alias{get_team_members}
\title{Get Team}
\usage{
get_team_members(id, ...)

get_team_boards(id, ...)
}
\arguments{
\item{id}{team ID, short name or URL}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a flat \code{data.frame} with team/organization-related data
}
\details{
Previously, teams were called "organizations", and the correct parent/child name in API calls remains "organization", "organizations".
}
\seealso{
\code{\link{trello_get}}
}

