% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trello.R
\name{get_trello}
\alias{get_trello}
\title{Get Data From Trello API}
\usage{
get_trello(url, token, query = list(limit = "1000"), paginate = FALSE)
}
\arguments{
\item{url}{url for the GET request, see \code{\link[httr]{GET}} for details}

\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}

\item{query}{url parameters that form the query, see \code{\link[httr]{GET}} for details}

\item{paginate}{logical whether paging should be used (if not, results will be limited to 1000 rows)}
}
\description{
This is a generic function that issues \code{\link[httr]{GET}} requests for trello API. It accepts JSON responses and converts them into a flat \code{data.frame} using \code{\link[jsonlite]{fromJSON}}. Other functions such as \code{\link{get_board_cards}} or \code{\link{get_card_comments}} are convenience wrappers for this function.
}
\examples{
# First, obtain a secure token to communicate with Trello API
token = get_token("your_key", "your_secret")

# Second, build an URL
id  = "id_of_a_given_board"
url = paste0("https://api.trello.com/1/boards/", id, "/cards")

# Third, store query parametrs in a list (see ?httr::GET for details)
query = list(filter = "open")

# Get all cards that are not archived
all_open = get_request(url, token, query)
}

