% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_trello_api}
\alias{get_trello_api}
\alias{put_trello_api}
\alias{post_trello_api}
\alias{delete_trello_api}
\title{GET/PUT/POST/DELETE trello api}
\usage{
get_trello_api(
  base_url,
  path,
  verbose = FALSE,
  content = TRUE,
  query = list(),
  key = NULL,
  token = NULL
)

put_trello_api(
  base_url,
  path,
  verbose = FALSE,
  content = TRUE,
  query = list(),
  body = NULL,
  encode = c("multipart", "form", "json", "raw"),
  key = NULL,
  token = NULL
)

post_trello_api(
  base_url,
  path,
  verbose = FALSE,
  content = TRUE,
  query = list(),
  body = NULL,
  encode = c("multipart", "form", "json", "raw"),
  key = NULL,
  token = NULL
)

delete_trello_api(base_url, path, verbose = TRUE, key = NULL, token = NULL)
}
\arguments{
\item{base_url}{string. the URL of the page to retrieve}

\item{path}{string. This string are passed on to \code{httr::modify_url()}.}

\item{verbose}{Boolean. Specify whether to give verbose output.}

\item{content}{Boolean. IF \code{TRUE}, Output of this function is passed to \code{httr::content()}.}

\item{query}{list. This list are passed on to \code{httr::modify_url()} with authentication information.}

\item{key, token}{string. The authentication information. See description.}

\item{body, encode}{These argument is passed \code{httr::POST()} or \code{httr::PUT()}. See these documents for more details.}
}
\value{
A list of API Responses.
If \code{content} is TRUE, Output is passed to \code{httr::content()}.
}
\description{
These are low-level functions that deal with CRUD.
The first word of the function name corresponds to each API method.
If the arguments key / token is NULL, it looks for Trello credentials on that local machine.
If the credentials are not registered on the running local machine, you need to run \code{set_trello_auth_infos ()} first.

Also, the key / token information is embedded in the query and API requested.
Be careful when handling the output of these functions.
}
