% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal.R
\name{get_url}
\alias{get_url}
\title{GET url from Trello API}
\usage{
get_url(url, token = NULL, response = "content", retry.times = 3,
  on.error = "error", handle = NULL, verbose = FALSE)
}
\arguments{
\item{url}{Url to get}

\item{token}{Secure token, see \code{\link{get_token}} for how to obtain it}

\item{response}{Can return \code{"content"} (a tibble, default), \code{"headers"}, \code{"status"} code or the complete \code{"response"}}

\item{retry.times}{Maximum number of requests to attempt}

\item{on.error}{Whether to stop, fail or message on error}

\item{handle}{The handle to use with this request (see \code{\link[httr]{RETRY}})}

\item{verbose}{Set to \code{TRUE} for verbose output}
}
\description{
GET url from Trello API
}
