% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoard_.R
\name{getBoard}
\alias{getBoard}
\title{Export Trello board}
\usage{
getBoard(path.json, archived = F, comments = F)
}
\arguments{
\item{path.json}{path to your JSON file}

\item{archived}{whether to export archived cards as well. Defaults to \code{FALSE}}

\item{comments}{whether to export comments as well. Defaults to \code{FALSE}}
}
\description{
This function takes JSON object obtained by Trello's native export function and converts it into an R data frame.
}
\examples{

# Download a Trello Board by pasting ".json" in the browser and save it as
# myboard.json in your working directory.

board = getBoard("myboard.json")

# To get archived cards and comments, do:

board = getBoard("myboard.json", archvied = T, comments = T)
}

