% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_s3.R
\name{is.trello_api}
\alias{is.trello_api}
\alias{is.trello_api}
\alias{is.boards_df}
\alias{is.cards_df}
\alias{is.lists_df}
\alias{is.actions_df}
\alias{is.comments_df}
\alias{is.labels_df}
\alias{is.members_df}
\alias{is.checklists_df}
\alias{is.fields_df}
\alias{is.teams_df}
\title{Trello API objects}
\usage{
is.trello_api(x)

is.boards_df(x)

is.cards_df(x)

is.lists_df(x)

is.actions_df(x)

is.comments_df(x)

is.labels_df(x)

is.members_df(x)

is.checklists_df(x)

is.fields_df(x)

is.teams_df(x)
}
\arguments{
\item{x}{Object to be tested}
}
\description{
Test for Trello API objects.
}
