% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal.R
\name{paginate}
\alias{paginate}
\title{GET multiple pages}
\usage{
paginate(url, token = NULL, response = "content", on.error = "error",
  handle = NULL, verbose = FALSE)
}
\arguments{
\item{url}{Url for the GET request}

\item{token}{Secure token, see \code{\link{get_token}} for how to obtain it}

\item{response}{Can return \code{"content"} (a tibble, default), \code{"headers"}, \code{"status"} code or the complete \code{"response"}}

\item{on.error}{Whether to stop, fail or message on error}

\item{handle}{The handle to use with this request (see \code{\link[httr]{RETRY}})}

\item{verbose}{Set to \code{TRUE} for verbose output}
}
\description{
GET multiple pages.
}
