% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model.R
\name{post_model}
\alias{post_model}
\title{POST data to Trello API}
\usage{
post_model(model, id = NULL, path = NULL, body = list(name = "New Card"),
  token, verbose = FALSE, ...)
}
\arguments{
\item{model}{Model}

\item{id}{Model id}

\item{path}{Path}

\item{body}{A named list of query paramters (will be passed as body)}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Wether to pass \code{verbose()} to \code{\link[httr]{POST}}}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
Issues \code{\link[httr]{POST}} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to post models.
}
