% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.trello_api}
\alias{print.actions_df}
\alias{print.cards_df}
\alias{print.labels_df}
\alias{print.trello_api}
\title{Printing Trello API objects}
\usage{
\method{print}{cards_df}(x)

\method{print}{actions_df}(x)

\method{print}{labels_df}(x)
}
\arguments{
\item{x}{object of class cards_df, members_df etc}
}
\description{
Print data.frame containing cards.
}

