% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_model.R
\name{search_model}
\alias{search_model}
\alias{search_cards}
\alias{search_actions}
\alias{search_boards}
\alias{search_members}
\alias{search_teams}
\title{Search models}
\usage{
search_model(string, modeltype = "all", boards = "mine",
  partial = FALSE, query = list(), ...)

search_cards(string, boards = "mine", partial = FALSE,
  query = list(), ...)

search_actions(string, boards = "mine", partial = FALSE,
  query = list(), ...)

search_boards(string, boards = "mine", partial = FALSE,
  query = list(), ...)

search_members(string, boards = "mine", partial = FALSE,
  query = list(), ...)

search_teams(string, boards = "mine", partial = FALSE,
  query = list(), ...)
}
\arguments{
\item{string}{Text to search for}

\item{modeltype}{Type of model to return; typically \code{"cards"}; defaults to \code{"all"}}

\item{boards}{Boards to limit the search to - defaults to \code{"mine"}}

\item{partial}{Should partial matching be used? Defaults to \code{FALSE}}

\item{query}{Name list of additional query parameters; consult \href{https://developers.trello.com/v1.0/reference}{Trello API reference}}

\item{...}{Additional arguments passed to \code{\link{get_model}}}
}
\description{
Returns a (nested) \code{data.frame} with search results.
}
\details{
Search results include one \code{data.frame} per each model type found. They
can be narrowed down by model type and other attributes (see \href{https://developers.trello.com/v1.0/reference}{Trello API reference})
}
\examples{
# Searches are only possible if authorized - a token is required:

\dontrun{

search_model("Anything with this text", token = mytoken)
search_cards("A card with this text", token = mytoken)

}
}
\seealso{
\code{\link{get_model}} \code{\link{get_token}}
}
