% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_board.R
\name{tidy_board}
\alias{tidy_board}
\title{Extract Table With Cards}
\usage{
tidy_board(board)
}
\arguments{
\item{board}{object of class board}
}
\description{
Use the 'trello.board' object acquired by the 'get_board' function to get a tidy data.frame with cards, their ids and other useful info.
}
\examples{
b = get_board(url, token)
btidy = tidy_board(b)
}

