% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_req.R
\name{trello_req}
\alias{trello_req}
\title{Build Requests For Trello API}
\usage{
trello_req(parent, child = NULL, filter = NULL)
}
\arguments{
\item{parent}{e.g. \code{"board"}, \code{"card"}, \code{"member"}}

\item{child}{e.g. \code{"cards"}, \code{"actions"}}

\item{filter}{character vector of length 1}
}
\description{
GET request function builder.
}
\seealso{
\code{\link{trello_get_token}}, \code{\link{trello_get}}
}

