% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_search.R
\name{trello_search}
\alias{trello_search}
\alias{trello_search_actions}
\alias{trello_search_boards}
\alias{trello_search_cards}
\alias{trello_search_members}
\alias{trello_search_teams}
\title{Search models}
\usage{
trello_search(string, ...)

trello_search_actions(string, ...)

trello_search_cards(string, ...)

trello_search_boards(string, ...)

trello_search_members(string, ...)

trello_search_teams(string, ...)
}
\arguments{
\item{string}{Text to search for}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a \code{list} with search results.
}
\examples{
# Searches are only possible if authorized - a token is required:

\dontrun{

trello_search("Anything with this text", my_token)
trello_search_cards("A card with this text", my_token)

}
}
\seealso{
\code{\link{trello_get}} \code{\link{trello_get_token}}
}

