% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trellor-package.R
\docType{package}
\name{trellor-package}
\alias{trellor}
\alias{trellor-package}
\title{trellor: 'Trello' 'API'}
\description{
A client to access the 'Trello' 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/trellor}
  \item \url{https://jonthegeek.github.io/trellor/}
  \item Report bugs at \url{https://github.com/jonthegeek/trellor/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
