% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_doc.R
\docType{package}
\name{trelloR}
\alias{trelloR}
\alias{trelloR-package}
\title{trelloR: R API for Trello}
\description{
Currently contains get_ and search_ functions for authorizing access to your
private boards and getting data. Summary and plotting functions will be
implemented in the future (watch \url{https://github.com/jchrom/trelloR}).
}

