% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_wrappers.R
\name{update_card}
\alias{update_card}
\title{Update card}
\usage{
update_card(card, body = NULL, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{body}{A named list of query parameters (will be passed as body with \code{encoding = "json"})}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{PUT}}}

\item{...}{Additional arguments passed to \code{\link[httr]{PUT}}}
}
\description{
Update name, description, assigned members and other fields.
}
