% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_wrappers.R
\name{update_checkitem}
\alias{update_checkitem}
\title{Update item}
\usage{
update_checkitem(card, checkitem, body = list(state = "complete"), ...)
}
\arguments{
\item{card}{Card id}

\item{checkitem}{Checklist item id}

\item{body}{A named list of query parameters (will be passed as body with \code{encoding = "json"}); defaults to \code{list(state = "complete")}, which checks the item out}

\item{...}{Additional arguments passed to \code{\link{put_model}}}
}
\description{
Update checklist item state, name, position and which chekclist it is in.
}
