% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{update_field}
\alias{update_field}
\title{Update custom field}
\usage{
update_field(id, body = list(name = "New name"), ...)
}
\arguments{
\item{id}{Board ID}

\item{body}{Named list with additional parameters}

\item{...}{Additional arguments passed to \code{\link{put_model}}}
}
\description{
Update custom field definition.
}
\seealso{
\code{\link{put_model}}
}
