% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_wrappers.R
\name{update_item}
\alias{update_item}
\title{Update item}
\usage{
update_item(card, item, body = list(state = "complete"), token,
  verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{item}{Checklist item id}

\item{body}{A named list of query parameters (will be passed as body with \code{encoding = "json"}); defaults to \code{list(state = "complete")}, which checks the item out}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{PUT}}}

\item{...}{Additional arguments passed to \code{\link[httr]{PUT}}}
}
\description{
Update checklist item state, name, position and which chekclist it is in.
}
