% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_board}
\alias{add_board}
\title{Add board}
\usage{
add_board(name, body = NULL, ...)
}
\arguments{
\item{name}{Name of the board.}

\item{body}{A named list of query paramters.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add board.
}
