% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_board}
\alias{add_board}
\title{Add board}
\usage{
add_board(name, body = NULL, ...)
}
\arguments{
\item{name}{Name of the board}

\item{body}{A named list of query paramters (will be passed as body)}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
POST board.
}
