% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_board}
\alias{add_board}
\title{Add board}
\usage{
add_board(name, body = NULL, ...)
}
\arguments{
\item{name}{Name of the board.}

\item{body}{A named list of query parameters.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add board.
}
\seealso{
Other functions to create resources: 
\code{\link{add_card}()},
\code{\link{add_checkitem}()},
\code{\link{add_checklist}()},
\code{\link{add_comment}()},
\code{\link{add_label}()},
\code{\link{add_list}()},
\code{\link{add_member}()},
\code{\link{create_resource}()}
}
\concept{functions to create resources}
