% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_card}
\alias{add_card}
\title{Add card}
\usage{
add_card(list, body = list(name = "New card"), token, verbose = FALSE, ...)
}
\arguments{
\item{list}{List id}

\item{body}{A named list of query paramters (will be passed as body)}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST card to a list.
}
