% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_card}
\alias{add_card}
\title{Add card}
\usage{
add_card(list, body = list(name = "New card"), ...)
}
\arguments{
\item{list}{List id.}

\item{body}{A named list of query parameters.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add card to a list.
}
\seealso{
Other functions to create resources: 
\code{\link{add_board}()},
\code{\link{add_card_attachment}()},
\code{\link{add_checkitem}()},
\code{\link{add_checklist}()},
\code{\link{add_comment}()},
\code{\link{add_label}()},
\code{\link{add_list}()},
\code{\link{add_member}()},
\code{\link{create_resource}()}
}
\concept{functions to create resources}
