% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_checkitem}
\alias{add_checkitem}
\title{Add checklist item}
\usage{
add_checkitem(checklist, name, checked = FALSE, position = "bottom", ...)
}
\arguments{
\item{checklist}{Checklist id}

\item{name}{Item name (text)}

\item{checked}{Whether item should be checked; defaults to \code{FALSE}}

\item{position}{Position in the checklist; defaults to \code{"bottom"}}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
POST item to a checklist.
}
