% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_checklist}
\alias{add_checklist}
\title{Add checklist}
\usage{
add_checklist(card, name, source = NULL, ...)
}
\arguments{
\item{card}{Card id.}

\item{name}{Checklist name.}

\item{source}{Items from this checklist id will be copied to the new one.
Defaults to \code{NULL}.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add checklist to a card.
}
