% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_comment}
\alias{add_comment}
\title{Add comment}
\usage{
add_comment(card, text, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{text}{Comment text}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST comment to a card.
}
