% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_comment}
\alias{add_comment}
\title{Add comment}
\usage{
add_comment(card, text, ...)
}
\arguments{
\item{card}{Card id.}

\item{text}{Comment text.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add comment to a card.
}
\seealso{
Other functions to create resources: 
\code{\link{add_board}()},
\code{\link{add_card}()},
\code{\link{add_checkitem}()},
\code{\link{add_checklist}()},
\code{\link{add_label}()},
\code{\link{add_list}()},
\code{\link{add_member}()},
\code{\link{create_resource}()}
}
\concept{functions to create resources}
