% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_list}
\alias{add_list}
\title{Add list}
\usage{
add_list(board, name, position = NULL, ...)
}
\arguments{
\item{board}{Board id}

\item{name}{List name}

\item{position}{List position (defaults to \code{NULL}); legal values are \code{"top"}, \code{"bottom"}}

\item{...}{Additional arguments passed to \code{\link{post_model}}}
}
\description{
POST list to a board.
}
