% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_list}
\alias{add_list}
\title{Add list}
\usage{
add_list(board, name, position = NULL, token, verbose = FALSE, ...)
}
\arguments{
\item{board}{Board id}

\item{name}{List name}

\item{position}{List position (defaults to \code{NULL}); legal values are \code{"top"}, \code{"bottom"}}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST list to a board.
}
