% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_member}
\alias{add_member}
\title{Add member}
\usage{
add_member(card, member, ...)
}
\arguments{
\item{card}{Card id.}

\item{member}{Member id.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add member to a card.
}
