% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_wrappers.R
\name{add_member}
\alias{add_member}
\title{Add member}
\usage{
add_member(card, member, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{member}{Member id}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{POST}} (for verbose output)}

\item{...}{Additional arguments passed to \code{\link[httr]{POST}}}
}
\description{
POST member to a card.
}
