% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wrappers.R
\name{add_member}
\alias{add_member}
\title{Add member}
\usage{
add_member(card, member, ...)
}
\arguments{
\item{card}{Card id.}

\item{member}{Member id.}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}.}
}
\description{
Add member to a card.
}
\seealso{
Other functions to create resources: 
\code{\link{add_board}()},
\code{\link{add_card}()},
\code{\link{add_checkitem}()},
\code{\link{add_checklist}()},
\code{\link{add_comment}()},
\code{\link{add_label}()},
\code{\link{add_list}()},
\code{\link{create_resource}()}
}
\concept{functions to create resources}
