% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_wrappers.R
\name{delete_card}
\alias{delete_card}
\title{Delete card}
\usage{
delete_card(card, ...)
}
\arguments{
\item{card}{Card id}

\item{...}{Additional arguments passed to \code{\link[=delete_resource]{delete_resource()}}}
}
\description{
DELETE card.
}
