% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_wrappers.R
\name{delete_checklist}
\alias{delete_checklist}
\title{Delete checklist}
\usage{
delete_checklist(checklist, ...)
}
\arguments{
\item{checklist}{Checklist id}

\item{...}{Additional arguments passed to \code{\link[=delete_resource]{delete_resource()}}}
}
\description{
DELETE checklist.
}
