% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{delete_field}
\alias{delete_field}
\title{Delete custom field}
\usage{
delete_field(id, ...)
}
\arguments{
\item{id}{Custom field ID (see \code{\link[=get_board_fields]{get_board_fields()}})}

\item{...}{Additional arguments passed to \code{\link[=create_resource]{create_resource()}}}
}
\description{
Delete custom field - this will remove it from all cards on the board.
}
\seealso{
Other functions to handle custom fields: 
\code{\link{add_field_option}()},
\code{\link{add_field}()},
\code{\link{delete_field_option}()},
\code{\link{update_card_field}()},
\code{\link{update_field}()}
}
\concept{functions to handle custom fields}
