% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_model.R
\name{delete_model}
\alias{delete_model}
\title{DELETE data via Trello API}
\usage{
delete_model(
  model,
  id = NULL,
  path = NULL,
  token = NULL,
  response = "content",
  on.error = c("stop", "warn", "message"),
  verbose = FALSE,
  encode = "json",
  handle = NULL
)
}
\arguments{
\item{model}{Model name, eg. \code{"card"}.}

\item{id}{Model id.}

\item{path}{Path.}

\item{token}{Secure token, see \link{get_token} (scope must include write
permissions).}

\item{response}{Can return \code{"content"} (the default), \code{"headers"}
or response object.}

\item{on.error}{Issues a warning, a message or an error on API error.}

\item{verbose}{Whether to pass \link[httr:verbose]{httr::verbose} to \link[httr:DELETE]{httr::DELETE}.}

\item{encode, handle}{Passed to \link[httr:DELETE]{httr::DELETE}.}
}
\description{
Issues \link[httr:DELETE]{httr::DELETE} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to DELETE requests.
}
\examples{

\dontrun{

# Get token with write access
token = get_token(yourkey, yoursecret, scope = c("read", "write"))

# Get board ID
url = "Your board URL"
bid = get_id_board(url, token)

# Get cards and extract ID of the first one
cid = get_board_cards(bid, token)$id[1]

# Delete it
delete_model(model = "card", id = cid, token = token)
}
}
