% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board}
\alias{get_board}
\title{Get Trello Board}
\usage{
get_board(url, token = NULL)
}
\arguments{
\item{url}{url of the Trello board, or a local file downloaded from Trello}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
Read in data either via API using oauth1 token or from a local file (which you can get by downloading data from Trello and saving as a text file)
}
\examples{
# Source credentials from non-shared location (important!)
source("mykeys.R")
token = get_token(key, secret)

You will be prompted to confirm the authorization in a browser. You will also be offered an option to store the authentication in your working directory, in a hidden '.httr-oauth' file (do NOT share it with anyone!).
}

