% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board}
\alias{get_board}
\alias{get_board_actions}
\alias{get_board_cards}
\alias{get_board_checklists}
\alias{get_board_comments}
\alias{get_board_labels}
\alias{get_board_lists}
\alias{get_board_members}
\alias{get_board_fields}
\alias{get_board_prefs}
\title{Get Board}
\usage{
get_board_actions(id, ...)

get_board_cards(id, ...)

get_board_checklists(id, ...)

get_board_comments(id, ...)

get_board_labels(id, ...)

get_board_lists(id, ...)

get_board_members(id, ...)

get_board_fields(id, ...)

get_board_prefs(id, ...)
}
\arguments{
\item{id}{Board ID}

\item{...}{Additional arguments passed to \code{\link[=get_resource]{get_resource()}}}
}
\description{
Returns a flat data frame with board-related data.
}
\examples{
\dontrun{
#Get cards with custom fields
cards = get_board_cards(board_id, query = list(customFieldItems = "true"))
}
}
\seealso{
\code{\link[=get_resource]{get_resource()}}
}
