% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R, R/get_card.R
\name{get_board}
\alias{get_board}
\alias{get_board_actions}
\alias{get_board_cards}
\alias{get_board_checklists}
\alias{get_board_comments}
\alias{get_board_labels}
\alias{get_board_lists}
\alias{get_board_members}
\alias{get_card_checklists}
\title{Get Board}
\usage{
get_board_actions(id, ...)

get_board_cards(id, ...)

get_board_checklists(id, ...)

get_board_comments(id, ...)

get_board_labels(id, ...)

get_board_lists(id, ...)

get_board_members(id, ...)

get_card_checklists(id, ...)
}
\arguments{
\item{id}{Board ID}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a flat \code{data.frame} with board-related data.
}
\seealso{
\code{\link{trello_get}}
}

