% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{get_board_cards}
\alias{get_board_cards}
\title{Get Board Cards}
\usage{
get_board_cards(id, token, filter = "all", paging = FALSE,
  simplify = TRUE)
}
\arguments{
\item{id}{id of the desired board}

\item{token}{previously generated token, see \code{\link{get_token}} for how to obtain it}

\item{filter}{whether to return "all" (the default), archived ("closed") or "open" cards}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{drop and rename some columns or leave it as is. Defaults to \code{TRUE}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with card-related data.
}
\examples{
cards = get_board_cards(id, token)
}
\seealso{
\code{\link{get_token}}, \code{\link{get_trello}}
}

