% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board_comments}
\alias{get_board_comments}
\title{Get Board Comments}
\usage{
get_board_comments(id, ...)
}
\arguments{
\item{id}{board id}

\item{...}{additional arguments passed to \code{\link{trello_get}}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with labels-related data.
}
\seealso{
\code{\link{trello_get}}
}

