% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board.R
\name{get_board_comments}
\alias{get_board_comments}
\title{Get Board Comments}
\usage{
get_board_comments(id, token = NULL, paging = FALSE, fix = TRUE)
}
\arguments{
\item{id}{board id}

\item{token}{previously generated token, see \code{\link{trello_get_token}} for info on how to obtain it}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{fix}{drop and rename some columns or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with labels-related data.
}
\seealso{
\code{\link{trello_get_token}}, \code{\link{trello_get}}
}

