% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{get_board_labels}
\alias{get_board_labels}
\title{Get Board Labels}
\usage{
get_board_labels(id, token, paginate = FALSE, simplify = TRUE)
}
\arguments{
\item{id}{id of the desired board}

\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}

\item{paginate}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{drop and rename some columns or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with labels-related data.
}
\examples{
labels = get_board_labels(id, token)
}

