% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board_requests.R
\name{get_board_lists}
\alias{get_board_lists}
\title{Get All Trello Board Lists}
\usage{
get_board_lists(boardid, token, filter = "all", paginate = FALSE,
  simplify = TRUE)
}
\arguments{
\item{boardid}{id of the desired board}

\item{token}{previously generated token (see ?get_token for help)}

\item{filter}{character whether to return only archived ("closed"), "open" or "all" (which is the default)}

\item{paginate}{logical whether to use pagination (for requests returning more than 1000 rows). Defaults to 'FALSE'}

\item{simplify}{logical drop and rename some columns to make the result simpler}
}
\description{
Returns a flat \code{data.frame} with all lists from a given Trello board.
}
\examples{
cards = get_board_lists(url, token)
}

