% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board.R
\name{get_board_lists}
\alias{get_board_lists}
\title{Get Board Lists}
\usage{
get_board_lists(id, token, filter = "all", paginate = FALSE,
  simplify = TRUE)
}
\arguments{
\item{id}{id of the desired board}

\item{token}{previously generated token, see \code{\link{get_token}} for how to obtain it}

\item{filter}{whether to return "all" (the default), archived ("closed") or "open" cards}

\item{paginate}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{drop and rename some columns or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Given a board ID, returns a flat \code{data.frame} with lists-related data.
}
\examples{
lists = get_board_lists(id, token)
}

