% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_board_requests.R
\name{get_board_members}
\alias{get_board_members}
\title{Get All Trello Board Memebers}
\usage{
get_board_members(boardid, token)
}
\arguments{
\item{boardid}{id of the desired board}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
Returns a flat \code{data.frame} with all members from a given Trello board.
}
\examples{
lists = get_board_members(url, token)
}

