% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{get_card_comments}
\alias{get_card_comments}
\title{Get All Trello Card Comments}
\usage{
get_card_comments(cardid, token)
}
\arguments{
\item{token}{previously generated token (see ?get_token for help)}

\item{boardid}{id of the desired board}
}
\description{
Returns a flat \code{data.frame} with all comments from a given Trello card.
}
\examples{
comments = get_card_comments(cardid, token)
}

