% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_card.R
\name{get_card_comments}
\alias{get_card_comments}
\title{Get Card Comments}
\usage{
get_card_comments(id, token, paging = FALSE, fix = TRUE)
}
\arguments{
\item{id}{card id}

\item{token}{previously generated token, see \code{\link{trello_get_token}} for how to obtain it}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{fix}{drop and rename some columns or leave it as is. Defaults to \code{TRUE}}
}
\description{
Given a card ID, returns a flat \code{data.frame} with comments-related data.
}
\seealso{
\code{\link{trello_get_token}}, \code{\link{trello_get}}
}

