% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{get_card_comments}
\alias{get_card_comments}
\title{Get Card Comments}
\usage{
get_card_comments(id, token, paginate = FALSE, simplify = TRUE)
}
\arguments{
\item{id}{id of the desired card}

\item{token}{previously generated token, see \code{\link{get_token}} for how to obtain it}

\item{paginate}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{drop and rename some columns or leave it as is. Defaults to \code{TRUE}}
}
\description{
Given a card ID, returns a flat \code{data.frame} with comments-related data.
}
\examples{
comments = get_card_comments(id, token)
}

