% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_card.R
\name{get_card_members}
\alias{get_card_members}
\title{Get Card Members}
\usage{
get_card_members(id, token, paging = FALSE, fix = TRUE)
}
\arguments{
\item{id}{card id}

\item{token}{previously generated token, see \code{\link{get_token}} for how to obtain it}

\item{paging}{whether paging should be used (necessary for requests returning more than 1000 rows). Defaults to \code{FALSE}}

\item{simplify}{drop and rename some columns or leave it as it is. Defaults to \code{TRUE}}
}
\description{
Given a card ID, returns a flat \code{data.frame} with members-related data.
}
\seealso{
\code{\link{get_token}}, \code{\link{get_trello}}
}

