% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_flat}
\alias{get_flat}
\title{GET url and return data.frame}
\usage{
get_flat(url, token = NULL, query = NULL)
}
\arguments{
\item{url}{Url to get}

\item{token}{Secure token}

\item{query}{Additional url parameters (defaults to NULL)}
}
\description{
GET url and return data.frame
}

