% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flat.R
\name{get_flat}
\alias{get_flat}
\title{GET url and return data.frame}
\usage{
get_flat(url, token = NULL, query = NULL, retry = 3, wait = 1.5)
}
\arguments{
\item{url}{Url to get}

\item{token}{Secure token, see \code{\link{get_token}} for how to obtain it}

\item{query}{Additional url parameters (defaults to NULL)}

\item{retry}{How many tries in case of server-side error}

\item{wait}{How long to sleep between retries (sec)}
}
\description{
GET url and return data.frame
}
