% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_get.R
\name{get_flat}
\alias{get_flat}
\title{GET url and return data.frame}
\usage{
get_flat(url, token = NULL, query = NULL)
}
\arguments{
\item{url}{url to get}

\item{token}{a secure token}

\item{query}{additional url parameters (defaults to NULL)}
}
\description{
GET url and return data.frame
}

