% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\alias{get_id_board}
\alias{get_id_card}
\title{Get model ID}
\usage{
get_id_board(url, token = NULL)

get_id_card(url, token = NULL)
}
\arguments{
\item{url}{Complete url, short url or just the url ID part of a Trello board}

\item{token}{Secure token, see \code{\link[=get_token]{get_token()}}}
}
\description{
Get ID of a resource.
}
\examples{
# Get Trello Development Roadmap board ID
url = "https://trello.com/b/nC8QJJoZ/trello-development-roadmap"
tdr_id = get_id_board(url)

# Also works:
url = "nC8QJJoZ"
tdr_id = get_id_board(url)
}
