% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_member_boards}
\alias{get_member_boards}
\title{Get Member's Boards}
\usage{
get_member_boards(id, ...)
}
\arguments{
\item{id}{member ID}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a flat \code{data.frame} with all boards associated with a member.
}
\seealso{
\code{\link{trello_get}}
}

