% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_member_boards}
\alias{get_member_boards}
\title{Get Member's Boards}
\usage{
get_member_boards(id, ...)
}
\arguments{
\item{id}{member ID or username}

\item{...}{Additional arguments passed to \code{\link[=get_resource]{get_resource()}}}
}
\description{
Returns a flat data frame with member-related data.
}
\seealso{
\code{\link[=get_resource]{get_resource()}}
}
