% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get Data From Trello API}
\usage{
get_model(parent = NULL, child = NULL, id = NULL, token = NULL,
  query = NULL, url = NULL, filter = NULL, limit = 1000,
  response = "content", paging = FALSE, bind.rows = TRUE,
  add.class = TRUE, on.error = "error", handle = NULL,
  verbose = FALSE)
}
\arguments{
\item{parent}{Parent resource (e.g. \code{"board"})}

\item{child}{Child resource (e.g. \code{"card"})}

\item{id}{Model ID}

\item{token}{Secure token, see \code{\link{get_token}} for how to obtain it; if NULL, it will check for a cached token}

\item{query}{Key-value pairs which form the query, see \code{\link[httr]{GET}} for details}

\item{url}{Url for the GET request, use instead of specifying \code{parent}, \code{id} and \code{child}; see \code{\link[httr]{GET}} for details}

\item{filter}{Defaults to \code{"all"} which fetches both open and archived cards or all action types}

\item{limit}{Defaults to \code{1000}; set to \code{0} to get everything}

\item{response}{Can be either "content" (a \code{\link[dplyr]{tbl}}) or an object of class \code{\link[httr]{response}}}

\item{paging}{Deprecated, use \code{limit = 0} instead}

\item{bind.rows}{Deprecated; will always try to bind rows unless \code{response} is not \code{"content"}}

\item{add.class}{Assign additional S3 class (defaults to \code{TRUE}). The additional class is currently not being used for anything, and might be dropped in the future.}

\item{on.error}{Whether to stop, fail or message on error}

\item{handle}{The handle to use with this request (see \code{\link[httr]{RETRY}})}

\item{verbose}{Set to \code{TRUE} for verbose output}
}
\description{
Issues \code{\link[httr]{GET}} requests for Trello API endpoints.
}
\details{
If the request fails, server error messages are reprinted on the console.

Only JSON responses are accepted. \code{\link[jsonlite]{fromJSON}} converts them into flat \code{data.frame}s or \code{list}s, while non-JSON type of response throws an error.

When \code{limit > 1000}, paging is used. The ID of the oldest result is retrieved from every page and supplied to the next request as the value of the \code{"before"} parameter. Paging will stop after reaching the number of results specified by \code{limit}, or when there are no more data to get.

\code{filter} and \code{limit} are query parameters that can be set individually; you could achieve the same result by using \code{query = list(filter = "filter_value", limit = "limit_value")}
}
\examples{

# No authorization is required to access public boards, so there is no need
# to create a token, just the board id:
url = "https://trello.com/b/Pw3EioMM/trellor-r-api-for-trello"
bid = get_id_board(url)

# Once we have the ID, we can use it to make specific queries:
labels = get_board_labels(bid)                # Get all labels
cards = get_board_cards(bid, filter = "all")  # Get all cards, incl. archived

# We can also call get_model() directly:
lists = get_model(parent = "board", child = "lists", id = bid)

# As with boards, cards can be queried for particular resources:
card10 = cards$id[10]
acts10 = get_card_actions(card10)    # Get all actions performed on that card

# To specify the number of results, use limit = number. If limit = 0, all
# results will be acquired eventually.

\dontrun{
acts_all = get_board_actions(bid, limit = 0)
}

# For private and team boards, you need a secure token:

\dontrun{
token = get_token("your_key", "your_secret")
cards_open = get_board_cards(board_id, token, filter = "open")
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}, \code{\link{get_token}}, \code{\link{get_id}}
}
