% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member.R
\name{get_my_boards}
\alias{get_my_boards}
\title{Get All Trello Member Boards}
\usage{
get_my_boards(token)
}
\arguments{
\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}
}
\description{
Returns a flat \code{data.frame} with all your boards.
}
\examples{
myboards = get_my_boards(token)
}

