% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_my_boards}
\alias{get_my_boards}
\title{Get Own Boards}
\usage{
get_my_boards(token, ...)
}
\arguments{
\item{token}{previously generated token, see \code{\link{trello_get_token}} for info on how to obtain it}

\item{...}{Additional arguments passed to \code{\link{trello_get}}}
}
\description{
Returns a flat \code{data.frame} with all your boards.
}
\seealso{
\code{\link{trello_get}}
}

