% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member.R
\name{get_my_boards}
\alias{get_my_boards}
\title{Get Own Boards}
\usage{
get_my_boards(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=get_resource]{get_resource()}}}
}
\description{
Returns a flat data frame, containing the authenticated user's boards.
}
\seealso{
\code{\link[=get_resource]{get_resource()}}
}
