% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Request Data From Trello}
\usage{
get_request(endpoint, id, query, token)
}
\arguments{
\item{endpoint}{character ("organization", "board", "card"...)}

\item{id}{id of the given endpoint}

\item{query}{url query, such as '?lists=open'}

\item{token}{previously generated token (see ?get_token for help)}
}
\description{
This is a generic function to perform query via trello API. Other function such as 'get_board_cards' or 'get_board_lists' are convenience wrappers for this function.
}
\examples{
all_cards = get_request("boards", 123456789, "cards", token)
}

