% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Request Data From Trello}
\usage{
get_request(url, token, query = list(limit = "1000"), paginate = FALSE)
}
\arguments{
\item{url}{url for the GET request, see \code{\link[httr]{GET}}}

\item{token}{previously generated token, see \code{\link{get_token}} for info on how to obtain it}

\item{query}{url query see \code{\link[httr]{GET}}}

\item{paginate}{logical whether pagination should be used (if not, results will be limited to 50 rows)}
}
\description{
This is a generic function to perform query via trello API. Other function such as 'get_board_cards' or 'get_board_lists' are convenience wrappers for this function.
}
\examples{
all_cards = get_request(, token)
}

