% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Authorize API access to Trello}
\usage{
get_token(key, secret)
}
\arguments{
\item{key}{key}

\item{secret}{secret}
}
\description{
To get data directly from Trello via API, get your 'key' and 'secret' here: https://trello.com/app-key.
}
\examples{
# Source credentials from non-shared location (important!)
source("mykeys.R")
token = get_token(key, secret)

# You will be prompted to confirm the authorization in a browser. You will
# also be offered an option to store the authentication in your working
# directory, in a hidden '.httr-oauth' file (do NOT share it with anyone!).
}

