% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get OAuth1.0 Token}
\usage{
get_token(key, secret, appname = "trello-app", scope = "read",
  expiration = "30days")
}
\arguments{
\item{key}{developer key}

\item{secret}{developer secret}

\item{appname}{optional app name, defaults to "trello-app"}

\item{scope}{defaults to \code{"read"}; allowed are \code{c("read", "write", "account")} or \code{"read,write,account"} if you prefer}

\item{expiration}{defaults to \code{"30days"}; also allowed \code{"1hour"}, \code{"1day"} and \code{"never"}}
}
\description{
Authorize access to Trello API.
}
\details{
To access private data, a secure token is required. In order to create it, you will need your developer credentials ('key' and 'secret'). These can be obtained visiting `https://trello.com/app-key` after logging in to Trello.

First time you create a token, you will be prompted to confirm the authorization in a browser (you only need to do this once). You will also be offered an option to store the authentication data in your working directory. Keep in mind you have to store your credentials in a \strong{secure, non-shared} location.

\code{\link{get_token}} uses authentication functions from \code{\link{httr}}.
}
\examples{
# Source credentials from non-shared location (important!)

\dontrun{

source("mykeys.R")
token = get_token(key, secret, scope = "read,write")
}
}
\seealso{
\code{\link[httr]{oauth_app}}, \code{\link[httr]{oauth_endpoint}}, \code{\link[httr]{oauth1.0_token}}
}
