% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_wrappers.R
\name{move_card}
\alias{move_card}
\title{Move card}
\usage{
move_card(card, list, token, verbose = FALSE, ...)
}
\arguments{
\item{card}{Card id}

\item{list}{List id}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{PUT}}}

\item{...}{Additional arguments passed to \code{\link[httr]{PUT}}}
}
\description{
Move card to another list.
}
