% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wrappers.R
\name{move_card}
\alias{move_card}
\title{Move card}
\usage{
move_card(card, list, ...)
}
\arguments{
\item{card}{Card id.}

\item{list}{List id.}

\item{...}{Additional arguments passed to \code{\link[=update_resource]{update_resource()}}.}
}
\description{
Move card to another list.
}
