% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_model.R
\name{post_model}
\alias{post_model}
\title{POST data to Trello API}
\usage{
post_model(
  model,
  id = NULL,
  path = NULL,
  body = list(name = "New"),
  token = NULL,
  response = c("content", "headers", "status", "response"),
  on.error = c("stop", "warn", "message"),
  verbose = FALSE,
  encode = "json",
  handle = NULL
)
}
\arguments{
\item{model}{Model name, eg. \code{"card"}.}

\item{id}{Model id.}

\item{path}{Path.}

\item{body}{A named list.}

\item{token}{Secure token, see \link{get_token} (scope must include write
permissions).}

\item{response}{Can return \code{"content"} (the default), \code{"headers"}
or response object.}

\item{on.error}{Issues a warning, a message or an error on API error.}

\item{verbose}{Whether to pass \link[httr:verbose]{httr::verbose} to \link[httr:POST]{httr::POST}.}

\item{encode, handle}{Passed to \link[httr:POST]{httr::POST}.}
}
\description{
Issues \link[httr:POST]{httr::POST} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to POST requests.
}
\examples{

\dontrun{

# Get token with write access
token = get_token(yourkey, yoursecret, scope = c("read", "write"))

# Get board ID
url = "Your board URL"
bid = get_id_board(url, token)

# Get lists on that board, extract ID of the first one
lid = get_board_lists(bid, token)$id[1]

# Content for the new card
payload = list(
  idList = lid,
  name = "A new card",
  desc = "#This card has been created by trelloR",
  pos = "bottom"
)

# Create card and store the response (to capture the ID
# of the newly created model)
r = post_model(model = "card", body = payload, token = token)

# Get ID of the new card
r$id
}
}
