% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.trello_api}
\alias{print.trello_api}
\alias{print.cards_df}
\alias{print.actions_df}
\alias{print.labels_df}
\alias{print.checklists_df}
\title{Printing Trello API objects}
\usage{
\method{print}{cards_df}(x, ...)

\method{print}{actions_df}(x, ...)

\method{print}{labels_df}(x, ...)

\method{print}{checklists_df}(x, ...)
}
\arguments{
\item{x}{Object of class \code{cards_df}, \code{actions_df} etc}

\item{...}{Further arguments passed to \code{\link[base]{print}}}
}
\description{
Print data.frame containing cards.
}
