% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_token.R
\name{print.Trello_API_token}
\alias{print.Trello_API_token}
\title{Print Trello API Tokens}
\usage{
\method{print}{Trello_API_token}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"Trello_API_token"}.}

\item{...}{Unused.}
}
\description{
Objects of class \code{"Trello_API_token"} posses some additional attributes to
make it easier to see the app name, its scope and expiration. If you want to
print the bare token object as generated by \verb{[httr::oauth1.0_token()]}, call
its own print method, eg: \code{my_token$print()}.
}
