% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_model.R
\name{put_model}
\alias{put_model}
\title{PUT data to Trello API}
\usage{
put_model(
  model,
  id = NULL,
  path = NULL,
  body = NULL,
  token = NULL,
  response = c("content", "headers", "status", "response"),
  on.error = c("stop", "warn", "message"),
  verbose = FALSE,
  encode = "json",
  handle = NULL
)
}
\arguments{
\item{model}{Model name, eg. \code{"card"}.}

\item{id}{Model id.}

\item{path}{Path.}

\item{body}{A named list.}

\item{token}{Secure token, see \link{get_token} (scope must include write
permissions).}

\item{response}{Can return \code{"content"} (the default), \code{"headers"}
or response object.}

\item{on.error}{Issues a warning, a message or an error on API error.}

\item{verbose}{Whether to pass \link[httr:verbose]{httr::verbose} to \link[httr:PUT]{httr::PUT}.}

\item{encode, handle}{Passed to \link[httr:PUT]{httr::PUT}.}
}
\description{
Issues \link[httr:PUT]{httr::PUT} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to POST requests.
}
\examples{

\dontrun{

# Get token with write access
token = get_token(yourkey, yoursecret, scope = c("read", "write"))

# Get board ID
url = "Your board URL"
bid = get_id_board(url, token)

# Get cards and extract ID of the first one
cid = get_board_cards(bid, token)$id[1]

# Content for the new card
payload = list(
  id = cid,
  name = "A new card name",
  desc = "Description - updated by trelloR",
  pos = "top" #put card on the top of a list
)

# Update card's name, descriptionand position
put_model(model = "card", id = cid, body = payload, token = token)
}
}
