% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_model.R
\name{put_model}
\alias{put_model}
\title{PUT data to Trello API}
\usage{
put_model(model, id = NULL, path = NULL, body = NULL, token,
  verbose = FALSE, response = "content", on.error = "warning", ...)
}
\arguments{
\item{model}{Model}

\item{id}{Model id}

\item{path}{Path}

\item{body}{A named list of query paramters (will be passed as body)}

\item{token}{Secure token, see \code{\link{get_token}} (scope must include write permissions)}

\item{verbose}{Whether to pass \code{verbose()} to \code{\link[httr]{PUT}}}

\item{response}{Can return \code{"content"} (default), \code{"headers"}, \code{"status"} code or the complete \code{"response"}}

\item{on.error}{Issues either \code{\link[base]{warning}} (default), \code{\link[base]{message}} or error (and \code{\link[base]{stop}}s)}

\item{...}{Additional arguments passed to \code{\link[httr]{PUT}}}
}
\description{
Issues \code{\link[httr]{PUT}} requests for Trello API endpoints.
}
\details{
See \href{https://developers.trello.com/v1.0/reference}{Trello API reference}
for more info about what arguments can be passed to PUT requests.
}
