% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_get.R
\name{trello_get}
\alias{trello_get}
\title{Get Data From Trello API}
\usage{
trello_get(url, token = NULL, query = NULL, paging = FALSE)
}
\arguments{
\item{url}{url for the GET request, see \code{\link[httr]{GET}} for details}

\item{token}{previously generated token, see \code{\link{trello_get_token}} for how to obtain it}

\item{query}{url parameters that form the query, see \code{\link[httr]{GET}} for details}

\item{paging}{logical whether paging should be used (if not, results will be limited to 1000 rows)}
}
\description{
Issues \code{\link[httr]{GET}} requests for Trello API endpoints.
}
\details{
It accepts JSON responses and uses \code{\link[jsonlite]{fromJSON}} to convert them into flat \code{data.frame}s. It hrows an error if non-JSON type of data is received. It also takes care of paging (if \code{paging = TRUE}), setting the query parameter \code{before} to the earliest date from the previous result, essentially getting all there is for the given request.

For unsuccessfull requests, server error messages are extracted from the response header and reprinted on the console.

Functions such as \code{\link{get_board_cards}} or \code{\link{get_card_comments}} are convenience wrappers for this function. They remove the need for specifying the query parameters and optionally simplify the result so as to give you more consistent and easy to work with data.
}
\examples{
# For accessing public boards you don't need authorization; this example uses
# the publicly available Trello Development Roadmap board (notice the .json
# suffix):
url = "https://trello.com/b/nC8QJJoZ/trello-development-roadmap.json"
tdb = trello_get(url)

# This gives you some useful content already, but you may want to do more
# specific queries. Let's start by getting the ID of the board:
bid = tdb$id

# We can now use this id to make specific queries using dedicated functions:

tdb_lists = get_board_lists(bid)   # Get lists
tdb_labels = get_board_labels(bid) # Get labels
tdb_cards = get_board_cards(bid)   # Get cards

# Having acquired the card-related data, we can now make queries about specific
# cards. As before, we start by getting the ID of the first card:
card1_id = tdb_cards$id[1]
card1_comm = get_card_comments(card1_id) # Get comments from the card

# To retrieve large results, a paging might be necessary:

\dontrun{
tdb_actions = get_board_actions(bid, filter = "commentCard", paging = TRUE)
}

# For private boards, you need a secure token to communicate with Trello API

\dontrun{
token = get_token("your_key", "your_secret")

# Get all cards that are not archived
all_open = get_request(url, token, query)
}
}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}, \code{\link{trello_get_token}}
}

