% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_get.R
\name{trello_get}
\alias{trello_get}
\title{Get Data From Trello API}
\usage{
trello_get(
  parent = NULL,
  child = NULL,
  id = NULL,
  token = NULL,
  query = NULL,
  url = NULL,
  filter = NULL,
  limit = 1000,
  paging = FALSE,
  bind.rows = TRUE
)
}
\arguments{
\item{parent, child, id, token, query, url, filter, limit, paging, bind.rows}{Use \code{\link{get_model}}.}
}
\description{
Deprecated. Use \code{\link{get_model}} instead.
}
