% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelloR-package.R
\docType{package}
\name{trelloR-package}
\alias{trelloR}
\alias{trelloR-package}
\title{trelloR: Access the Trello API}
\description{
An R client for the Trello API. Supports free-tier features such as access
to private boards, creating and updating cards and other resources, and downloading
data in a structured way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jchrom/trelloR}
  \item \url{http://jchrom.github.io/trelloR/}
  \item Report bugs at \url{https://github.com/jchrom/trelloR/issues}
}

}
\author{
\strong{Maintainer}: Jakub Chromec \email{chromec.jakub@gmail.com}

}
\keyword{internal}
