% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_doc.R
\docType{package}
\name{trelloR}
\alias{trelloR}
\title{trelloR: R API for Trello}
\description{
Authorize your access to Trello from R and get data from boards, cards, lists
and other places. Create and update and remove cards and boards, add labels,
assign team members, add comments and more.
}
