% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelloR.R
\docType{package}
\name{trelloR}
\alias{trelloR}
\title{trelloR: R API for Trello}
\description{
Access the Trello API. Free-tier features are fully supported, including
authentication and administration of private boards, teams, creating
and updating cards and other resources, and downloading data in a structured
way.
}
