% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_wrappers.R
\name{update_card}
\alias{update_card}
\title{Update card}
\usage{
update_card(card, body = NULL, ...)
}
\arguments{
\item{card}{Card id}

\item{body}{A named list of query parameters (will be passed as body with \code{encoding = "json"})}

\item{...}{Additional arguments passed to \code{\link{put_model}}}
}
\description{
Update name, description, assigned members and other fields.
}
