% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wrappers.R
\name{update_card}
\alias{update_card}
\title{Update card}
\usage{
update_card(card, body = NULL, ...)
}
\arguments{
\item{card}{Card id}

\item{body}{A named list of query parameters.}

\item{...}{Additional arguments passed to \code{\link[=update_resource]{update_resource()}}.}
}
\description{
Update name, description, assigned members and other fields.
}
