% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wrappers.R
\name{update_card_labels}
\alias{update_card_labels}
\title{Update card labels}
\usage{
update_card_labels(card, labels, ...)
}
\arguments{
\item{card}{Card id.}

\item{labels}{A character vector of one or more label id.}

\item{...}{Additional arguments passed to \code{\link[=update_resource]{update_resource()}}.}
}
\description{
Replace currently assigned labels.
}
