% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wrappers.R
\name{update_card_members}
\alias{update_card_members}
\title{Update card members}
\usage{
update_card_members(card, members, ...)
}
\arguments{
\item{card}{Card id.}

\item{members}{A character vector of one or more member id.}

\item{...}{Additional arguments passed to \code{\link[=update_resource]{update_resource()}}.}
}
\description{
Replace currently assigned members.
}
