% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_wrappers.R
\name{update_checkitem}
\alias{update_checkitem}
\title{Update item}
\usage{
update_checkitem(card, checkitem, body = list(state = "complete"), ...)
}
\arguments{
\item{card}{Card id.}

\item{checkitem}{Checklist item id.}

\item{body}{A named list of query parameters. Defaults to
\code{list(state = "complete")}, which makes the item complete.}

\item{...}{Additional arguments passed to \code{\link[=update_resource]{update_resource()}}.}
}
\description{
Update checklist item state, name, position and which checklist it is in.
}
